feat: Implement RemoteFile struct for loading firebase files
